Move testing only functions in providercache to their own file
